Refactored JournalEntry to use long
Using long as ids rather than ints as well as replacing journalId with
long rather than journal object.

+++ src/main/java/com/journal/entities/Journal.java
@@ -18,7 +18,7 @@ public class Journal {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name="journal_id")
-    private int journalId;
+    private long journalId;
 
     @Column(name="journal_name")
     private String journalName;
